refactor: split snapshot.py into uec/ module
